pyproject.toml
[project]
name = "wasyagana-tv"
version = "0.1.0"
description = "Add your description here"
readme = "README.rst"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client~=2.142.0",
    "google-auth-oauthlib~=1.2.1",
    "google-auth-httplib2~=0.2.0",
    "pydantic~=2.8.2",
    "python-dateutil~=2.9.0.post0",
]

[project.scripts]
wasyagana-fetch-videos = "wasyagana_tv.cli.fetch_videos:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
